feat: 경건님 editor코드와 채팅 병합
