+++ app/ide/components/chat/ChatPanel.tsx
@@ -1,6 +1,8 @@
 "use client";
 
 import { useChat } from "@/app/ide/hooks/useChat";
+import MessageList from "./MessageList";
+import MessageInput from "./MessageInput";
 
 interface ChatPanelProps {
   containerId: string;
@@ -11,24 +13,13 @@ export default function ChatPanel({ containerId }: ChatPanelProps) {
 
   return (
     <div className="h-full flex flex-col p-4">
-      <div className="flex-1 overflow-y-auto mb-2">
-        {messages.map((m) => (
-          <div key={m.id}>
-            {m.sender}: {m.content}
-          </div>
-        ))}
-      </div>
+      <MessageList messages={messages} />
 
-      <div className="flex gap-2">
-        <input
-          className="border flex-1 px-2"
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-        />
-        <button className="border px-2" onClick={sendMessage}>
-          전송
-        </button>
-      </div>
+      <MessageInput
+        value={input ?? ""}
+        onChange={setInput}
+        onSend={sendMessage}
+      />
     </div>
   );
 }
