--- app/ide/components/editor/Editor.tsx
@@ -1,7 +1,0 @@
-export default function Editor() {
-  return (
-    <div className="h-full w-full flex items-center justify-center text-gray-400">
-      Editor Area
-    </div>
-  );
-}
