+++ app/ide/components/editor/Timer.tsx
