+++ app/ide/components/filetree/FileSidebar.tsx
