+++ app/ide/components/terminal/TerminalPanel.tsx
